backend/requirements-simple.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
itsdangerous==2.1.2